Crate actor

Modules

account

cron

init

market

miner

multisig

network

paych

power

reward

system

verifreg

Structs

ACCOUNT_ACTOR_CODE_ID

ActorState
State of all actor implementations

BURNT_FUNDS_ACTOR_ADDR

BalanceTable
Balance table which handles getting and updating token balances specifically

CALLER_TYPES_SIGNABLE

CRON_ACTOR_ADDR

CRON_ACTOR_CODE_ID

INIT_ACTOR_ADDR

INIT_ACTOR_CODE_ID

MARKET_ACTOR_CODE_ID

MINER_ACTOR_CODE_ID

MULTISIG_ACTOR_CODE_ID

Multimap
Multimap stores multiple values per key in a Hamt of Amts. The order of insertion of values for each key is retained.

OptionalEpoch

PAYCH_ACTOR_CODE_ID

POWER_ACTOR_CODE_ID

REWARD_ACTOR_ADDR

REWARD_ACTOR_CODE_ID

STORAGE_MARKET_ACTOR_ADDR

STORAGE_POWER_ACTOR_ADDR

SYSTEM_ACTOR_ADDR

SYSTEM_ACTOR_CODE_ID

Serialized
Serialized bytes to be used as parameters into actor methods

Set
Set is a Hamt with empty values for the purpose of acting as a hash set.

SetMultimap
SetMultimap is a hamt with values that are also a hamt but are of the set variant. This allows hash sets to be indexable by an address.

VERIFIED_ACTOR_CODE_ID

VERIFIED_REGISTRY_ACTOR_ADDR

Constants

FIRST_NON_SINGLETON_ADDR
Defines first available ID address after builtin actors

Functions

is_principal

parse_uint_key

u64_key

Type Definitions

DealID
Deal identifier used in market and miner actors