Struct actor::init::State

pub struct State {
    pub address_map: Cid,
    pub next_id: ActorID,
    pub network_name: String,
}

State is reponsible for creating

Fields
address_map: Cidnext_id: ActorIDnetwork_name: String
Implementations

impl State

pub fn new(address_map: Cid, network_name: String) -> Self

pub fn map_address_to_new_id<BS: BlockStore>(
&mut self,
store: &BS,
addr: &Address
) -> Result<Address, HamtError>

Allocates a new ID address and stores a mapping of the argument address to it. Returns the newly-allocated address.

pub fn resolve_address<BS: BlockStore>(
&self,
store: &BS,
addr: &Address
) -> Result<Address, String>

ResolveAddress resolves an address to an ID-address, if possible. If the provided address is an ID address, it is returned as-is. This means that ID-addresses (which should only appear as values, not keys) and singleton actor addresses pass through unchanged.

Post-condition: all addresses succesfully returned by this method satisfy addr.protocol() == Protocol::ID.