Module actor::market

Structs

Actor
Market Actor

ClientDealProposal
ClientDealProposal is a DealProposal signed by a client

ComputeDataCommitmentParams

DealProposal
Note: Deal Collateral is only released and returned to clients and miners when the storage deal stops counting towards power. In the current iteration, it will be released when the sector containing the storage deals expires, even though some storage deals can expire earlier than the sector does. Collaterals are denominated in PerEpoch to incur a cost for self dealing or minimal deals that last for a long time. Note: ClientCollateralPerEpoch may not be needed and removed pending future confirmation. There will be a Minimum value for both client and provider deal collateral.

DealState

HandleExpiredDealsParams

OnMinerSectorsTerminateParams

PublishStorageDealsParams

PublishStorageDealsReturn

State
Market actor state

VerifyDealsOnSectorProveCommitParams

VerifyDealsOnSectorProveCommitReturn

WithdrawBalanceParams

Enums

Method
Market actor methods available