Function actor::miner::assign_new_sectors

pub fn assign_new_sectors(
    deadlines: &mut Deadlines, 
    partition_size: usize, 
    new_sectors: &[u64], 
    _seed: Randomness
) -> Result<(), String>

Assigns a sequence of sector numbers to deadlines by:

filling any non-full partitions, in round-robin order across the deadlines

repeatedly adding a new partition to the deadline with the fewest partitions When multiple partitions share the minimal sector count, one is chosen at random (from a seed).