Struct actor::miner::State

pub struct State {
    pub info: MinerInfo,
    pub pre_commit_deposit: TokenAmount,
    pub locked_funds: TokenAmount,
    pub pre_committed_sectors: Cid,
    pub sectors: Cid,
    pub proving_period_start: ChainEpoch,
    pub new_sectors: BitField,
    pub sector_expirations: Cid,
    pub deadlines: Cid,
    pub faults: BitField,
    pub fault_epochs: Cid,
    pub recoveries: BitField,
    pub post_submissions: BitField,
    pub next_deadline_to_process_faults: usize,
    // some fields omitted
}

Fields
info: MinerInfo

Contains static info about this miner

pre_commit_deposit: TokenAmount

Total funds locked as pre_commit_deposit

locked_funds: TokenAmount

Total unvested funds locked as pledge collateral

pre_committed_sectors: Cid

Sectors that have been pre-committed but not yet proven. Map, HAMT<SectorNumber, SectorPreCommitOnChainInfo>

sectors: Cid

Sectors this miner has committed Array, AMT

proving_period_start: ChainEpoch

The first epoch in this miner's current proving period. This is the first epoch in which a PoSt for a partition at the miner's first deadline may arrive. Alternatively, it is after the last epoch at which a PoSt for the previous window is valid. Always greater than zero, his may be greater than the current epoch for genesis miners in the first WPoStProvingPeriod epochs of the chain; the epochs before the first proving period starts are exempt from Window PoSt requirements. Updated at the end of every period by a power actor cron event.

new_sectors: BitField

Sector numbers prove-committed since period start, to be added to Deadlines at next proving period boundary.

sector_expirations: Cid

Sector numbers indexed by expiry epoch (which are on proving period boundaries). Invariant: Keys(Sectors) == union(SectorExpirations.Values()) Array, AMT[ChainEpoch]Bitfield

deadlines: Cid

The sector numbers due for PoSt at each deadline in the current proving period, frozen at period start. New sectors are added and expired ones removed at proving period boundary. Faults are not subtracted from this in state, but on the fly.

faults: BitField

All currently known faulty sectors, mutated eagerly. These sectors are exempt from inclusion in PoSt.

fault_epochs: Cid

Faulty sector numbers indexed by the start epoch of the proving period in which detected. Used to track fault durations for eventual sector termination. At most 14 entries, b/c sectors faulty longer expire. Invariant: Faults == union(FaultEpochs.Values()) AMT[ChainEpoch]Bitfield

recoveries: BitField

Faulty sectors that will recover when next included in a valid PoSt. Invariant: Recoveries ⊆ Faults.

post_submissions: BitField

Records successful PoSt submission in the current proving period by partition number. The presence of a partition number indicates on-time PoSt received.

next_deadline_to_process_faults: usize

The index of the next deadline for which faults should been detected and processed (after it's closed). The proving period cron handler will always reset this to 0, for the subsequent period. Eager fault detection processing on fault/recovery declarations or PoSt may set a smaller number, indicating partial progress, from which subsequent processing should continue. In the range [0, WPoStProvingPeriodDeadlines).

Implementations

impl State

pub fn new(
empty_arr: Cid,
empty_map: Cid,
empty_deadlines: Cid,
owner: Address,
worker: Address,
peer_id: Vec<u8>,
multi_address: Vec<u8>,
seal_proof_type: RegisteredSealProof,
period_start: ChainEpoch
) -> Result<Self, String>

pub fn get_worker(&self) -> &Address

Returns worker address

pub fn get_sector_size(&self) -> &SectorSize

Returns sector size

pub fn deadline_info(&self, current_epoch: ChainEpoch) -> DeadlineInfo

Returns deadline calculations for the current proving period.

pub fn sector_count<BS: BlockStore>(&self, store: &BS) -> Result<u64, AmtError>

pub fn get_max_allowed_faults<BS: BlockStore>(
&self,
store: &BS
) -> Result<u64, AmtError>

pub fn put_precommitted_sector<BS: BlockStore>(
&mut self,
store: &BS,
info: SectorPreCommitOnChainInfo
) -> Result<(), HamtError>

pub fn get_precommitted_sector<BS: BlockStore>(
&self,
store: &BS,
sector_num: SectorNumber
) -> Result<Option<SectorPreCommitOnChainInfo>, HamtError>

pub fn delete_precommitted_sector<BS: BlockStore>(
&mut self,
store: &BS,
sector_num: SectorNumber
) -> Result<(), HamtError>

pub fn has_sector_number<BS: BlockStore>(
&self,
store: &BS,
sector_num: SectorNumber
) -> Result<bool, AmtError>

pub fn put_sector<BS: BlockStore>(
&mut self,
store: &BS,
sector: SectorOnChainInfo
) -> Result<(), AmtError>

pub fn get_sector<BS: BlockStore>(
&self,
store: &BS,
sector_num: SectorNumber
) -> Result<Option<SectorOnChainInfo>, AmtError>

pub fn delete_sector<BS: BlockStore>(
&mut self,
store: &BS,
sector_nos: &mut BitField
) -> Result<(), AmtError>

pub fn for_each_sector<BS: BlockStore, F>(
&self,
store: &BS,
f: F
) -> Result<(), String> where
F: FnMut(&SectorOnChainInfo) -> Result<(), String>,

pub fn add_new_sectors(
&mut self,
sector_nos: &[SectorNumber]
) -> Result<(), String>

Adds some sector numbers to the new sectors bitfield.

pub fn remove_new_sectors(
&mut self,
sector_nos: &BitField
) -> Result<(), String>

Removes some sector numbers from the new sectors bitfield, if present.

pub fn get_sector_expirations<BS: BlockStore>(
&self,
store: &BS,
expiry: ChainEpoch
) -> Result<BitField, String>

Gets the sector numbers expiring at some epoch.

pub fn for_each_sector_expiration<BS: BlockStore, F>(
&self,
store: &BS,
f: F
) -> Result<(), String> where
F: FnMut(ChainEpoch, &BitField) -> Result<(), String>,

Iterates sector expiration groups in order. Note that the sectors bitfield provided to the callback is not safe to store.

pub fn add_sector_expirations<BS: BlockStore>(
&mut self,
store: &BS,
expiry: ChainEpoch,
sectors: &[u64]
) -> Result<(), String>

Adds some sector numbers to the set expiring at an epoch. The sector numbers are given as uint64s to avoid pointless conversions.

pub fn remove_sector_expirations<BS: BlockStore>(
&mut self,
store: &BS,
expiry: ChainEpoch,
sectors: &[u64]
) -> Result<(), String>

Removes some sector numbers from the set expiring at an epoch.

pub fn clear_sector_expirations<BS: BlockStore>(
&mut self,
store: &BS,
expirations: &[ChainEpoch]
) -> Result<(), String>

Removes all sector numbers from the set expiring some epochs.

pub fn add_faults<BS: BlockStore>(
&mut self,
store: &BS,
sector_nos: &mut BitField,
fault_epoch: ChainEpoch
) -> Result<(), String>

Adds sectors numbers to faults and fault epochs.

pub fn remove_faults<BS: BlockStore>(
&mut self,
store: &BS,
sector_nos: &mut BitField
) -> Result<(), String>

Removes sector numbers from faults and fault epochs, if present.

pub fn for_each_fault_epoch<BS: BlockStore, F>(
&self,
store: &BS,
f: F
) -> Result<(), String> where
F: FnMut(ChainEpoch, &BitField) -> Result<(), String>,

Iterates faults by declaration epoch, in order.

pub fn clear_fault_epochs<BS: BlockStore>(
&mut self,
store: &BS,
epochs: &[ChainEpoch]
) -> Result<(), String>

pub fn add_recoveries(
&mut self,
sector_nos: &mut BitField
) -> Result<(), String>

Adds sectors to recoveries.

pub fn remove_recoveries(
&mut self,
sector_nos: &mut BitField
) -> Result<(), String>

Removes sectors from recoveries, if present.

pub fn load_sector_infos<BS: BlockStore>(
&self,
store: &BS,
sectors: &mut BitField
) -> Result<Vec<SectorOnChainInfo>, String>

Loads sector info for a sequence of sectors.

pub fn load_sector_infos_for_proof<BS: BlockStore>(
&mut self,
store: &BS,
proven_sectors: BitField
) -> Result<(Vec<SectorOnChainInfo>, BitField), String>

Loads info for a set of sectors to be proven. If any of the sectors are declared faulty and not to be recovered, info for the first non-faulty sector is substituted instead. If any of the sectors are declared recovered, they are returned from this method.

pub fn add_post_submissions(
&mut self,
partition_nos: BitField
) -> Result<(), String>

Adds partition numbers to the set of PoSt submissions

pub fn clear_post_submissions(&mut self) -> Result<(), String>

Removes all PoSt submissions

pub fn load_deadlines<BS: BlockStore>(
&self,
store: &BS
) -> Result<Deadlines, String>

pub fn save_deadlines<BS: BlockStore>(
&mut self,
store: &BS,
deadlines: Deadlines
) -> Result<(), String>

pub fn add_pre_commit_deposit(&mut self, amount: &TokenAmount)

pub fn subtract_pre_commit_deposit(&mut self, amount: &TokenAmount)

pub fn add_locked_funds<BS: BlockStore>(
&mut self,
store: &BS,
current_epoch: ChainEpoch,
vesting_sum: &TokenAmount,
spec: VestSpec
) -> Result<(), AmtError>

pub fn unlock_unvested_funds<BS: BlockStore>(
&mut self,
store: &BS,
current_epoch: ChainEpoch,
target: TokenAmount
) -> Result<TokenAmount, String>

Unlocks an amount of funds that have not yet vested, if possible. The soonest-vesting entries are unlocked first. Returns the amount actually unlocked.

pub fn unlock_vested_funds<BS: BlockStore>(
&mut self,
store: &BS,
current_epoch: ChainEpoch
) -> Result<TokenAmount, String>

Unlocks all vesting funds that have vested before the provided epoch. Returns the amount unlocked.

pub fn check_vested_funds<BS: BlockStore>(
&self,
store: &BS,
current_epoch: ChainEpoch
) -> Result<TokenAmount, String>

CheckVestedFunds returns the amount of vested funds that have vested before the provided epoch.

pub fn get_available_balance(&self, actor_balance: &TokenAmount) -> TokenAmount

pub fn assert_balance_invariants(&self, balance: &TokenAmount)