Module actor::power

Structs

Actor
Storage Power Actor

AddBalanceParams

CONSENSUS_MINER_MIN_POWER
Minimum power of an individual miner to meet the threshold for leader election.

Claim

CreateMinerParams

CreateMinerReturn

CronEvent

DeleteMinerParams

EPOCH_TOTAL_EXPECTED_REWARD
Total expected block reward per epoch (per-winner reward * expected winners), as input to pledge requirement.

EnrollCronEventParams

OnMinerWindowedPoStFailureParams

OnSectorModifyWeightDescParams

OnSectorProveCommitParams

OnSectorTemporaryFaultEffectiveBeginParams

OnSectorTemporaryFaultEffectiveEndParams

OnSectorTerminateParams

PLEDGE_FACTOR
Multiplier on sector pledge requirement.

ReportConsensusFaultParams

SectorStorageWeightDesc

State
Storage power actor state

WithdrawBalanceParams

Enums

Method
Storage power actor methods available

Constants

CONSENSUS_FAULT_REPORTING_WINDOW
Maximum age of a block header used as proof of a consensus fault to appear in the chain.

CONSENSUS_MINER_MIN_MINERS
Minimum number of registered miners for the minimum miner size limit to effectively limit consensus power.

SECTOR_TERMINATION_EXPIRED
Implicit termination after all deals expire

SECTOR_TERMINATION_MANUAL
Unscheduled explicit termination by the miner

WINDOWED_POST_CHALLENGE_DURATION
The time a miner has to respond to a surprise PoSt challenge.

WINDOWED_POST_FAILURE_LIMIT
The number of consecutive failures to meet a surprise PoSt challenge before a miner is terminated.

Functions

consensus_power_for_weight

pledge_for_weight

Type Definitions

SectorTermination