Crate actor

Structs

AccountActorCode

AccountActorState
AccountActorState includes the address for the actor

ActorID
Identifier for Actors, includes builtin and initialized actors

ActorState
State of all actor implementations

CronActorCode

CronActorState
CronActorState has no internal state

CronTableEntry

InitActorCode

InitActorState
InitActorState is reponsible for creating

Reward

RewardActorCode

RewardActorState
RewardActorState has no internal state

StoragePowerActorCode

StoragePowerActorState
State of storage power actor

Enums

AccountMethod

CodeID
CodeID is the reference to the code which is attached to the Actor state. There are builtin IDs and the option for custom code with a Cid

CronMethod

InitMethod

RewardMethod

StoragePowerMethod
Method definitions for Storage Power Actor

Traits

Actor
Actor trait which defines the common functionality of system Actors